requirements.txt
# Data Manipulation
pandas==2.2.0
openpyxl==3.1.2

# Database Connectivity
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Environment & Utilities
python-dotenv==1.0.0

# Analytics & Math
scipy==1.12.0
numpy==1.26.3